Run dep again with updated version
As seen on travis, there is an error with with old dep versions when
using packages with uppercase.

With version v0.3.2-211-g015a7570 that was used, there was an
inconsistency between import paths working on case insensitive systems
(mac OS, Windows) but failing on linux (c.f.
https://travis-ci.org/schibsted/docker-postgres-tests/jobs/357465804)

This commit performs a new run of dep using the latest version that
seems to solve the issue

Signed-off-by: Thibault Jamet <tjamet@users.noreply.github.com>